pyproject.toml
[tool.poetry]
name = "snowpycdk"
version = "0.1.0"
description = ""
authors = ["Amine Amarzouk <amine.amarzouk@servier.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
airbyte-cdk = "^6.32.0"
snowflake-connector-python = "^3.13.2"
pyarrow = "^19.0.0"
google-cloud-bigquery = "^3.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
